fixed main and tests. Properties now uses Header/Body model
While examining the files it seemed that a flag (null byte) was always present in a property.

It also appeared to be the start of the data (correlating to the property header's size field)

But there was extra data before (and some extra after) in certain cases

While extra data after is still unclear (and likely relates to the WorldObject instead of properties directly), extra data before is clearly tied to the property, and often describes it. This works rather nicely for most types; the exception being StructArray.